fix: Ide api 및 기능 구현 일부 완료, User Long 타입으로 변경

+++ backend/src/main/java/com/goorm/goormfriends/api/service/UserService.java
@@ -3,7 +3,7 @@ package com.goorm.goormfriends.api.service;
 import com.goorm.goormfriends.api.dto.response.LoginResponse;
 
 public interface UserService {
-    String findUserEmailByUserId(int userId) throws Exception;
+    String findUserEmailByUserId(Long userId) throws Exception;
 
     String oauthLogin(String email) throws Exception;
